Fix: Prediction Candidate Is Duplicated
After selecting a prediction candidate, the candidate is stored in the user dictionary. Inputting the same characters again, both the candidate stored in the user dictionary and the same prediction candidate appears. However, they are not deduplicated since the former has incorrect pronunciation (which appears to be correct with dictionary-lookup-filter), which differs from the pronunciation of the latter.